requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyMuPDF==1.23.25
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.52.2
SQLAlchemy==2.0.35
requests==2.32.3
ebooklib==0.18
beautifulsoup4==4.12.3
websocket-client==1.7.0
websockets==13.1
volcengine-python-sdk[ark]

